+++ src/types/quiz.type.ts
@@ -1,3 +1,5 @@
+import { QuizType } from '.';
+
 export interface QuizOptionType {
   type: string;
   amount: string;
@@ -13,4 +15,13 @@ export interface QuizCreationByFileType {
 export interface QuizCreationByTextType {
   option: QuizOptionType;
   text: string;
+}
+
+export interface CategoryQuizItemsType {
+  categorizedProblemId: number;
+  problemGeneratedBy: string;
+  problemType: QuizType;
+  problemName: string;
+  createTime: string;
+  updateTime: string;
 }
